refactor(Legend): Changes based on Feedback from @vogelino on #287
