+++ .vscode/settings.json
@@ -1,21 +1,22 @@
 {
-    "workbench.colorCustomizations": {
-        "activityBar.background": "#8a2ca2",
-        "activityBar.activeBackground": "#8a2ca2",
-        "activityBar.activeBorder": "#af9530",
-        "activityBar.foreground": "#e7e7e7",
-        "activityBar.inactiveForeground": "#e7e7e799",
-        "activityBarBadge.background": "#af9530",
-        "activityBarBadge.foreground": "#15202b",
-        "titleBar.activeBackground": "#68217a",
-        "titleBar.inactiveBackground": "#68217a99",
-        "titleBar.activeForeground": "#e7e7e7",
-        "titleBar.inactiveForeground": "#e7e7e799",
-        "statusBar.background": "#68217a",
-        "statusBarItem.hoverBackground": "#8a2ca2",
-        "statusBar.foreground": "#e7e7e7"
-    },
-    "peacock.color": "#68217A",
-    "typescript.tsdk": "node_modules/typescript/lib",
-    "eslint.enable": true
+  "workbench.colorCustomizations": {
+    "activityBar.background": "#8a2ca2",
+    "activityBar.activeBackground": "#8a2ca2",
+    "activityBar.activeBorder": "#af9530",
+    "activityBar.foreground": "#e7e7e7",
+    "activityBar.inactiveForeground": "#e7e7e799",
+    "activityBarBadge.background": "#af9530",
+    "activityBarBadge.foreground": "#15202b",
+    "titleBar.activeBackground": "#68217a",
+    "titleBar.inactiveBackground": "#68217a99",
+    "titleBar.activeForeground": "#e7e7e7",
+    "titleBar.inactiveForeground": "#e7e7e799",
+    "statusBar.background": "#68217a",
+    "statusBarItem.hoverBackground": "#8a2ca2",
+    "statusBar.foreground": "#e7e7e7"
+  },
+  "peacock.color": "#68217A",
+  "typescript.tsdk": "node_modules/typescript/lib",
+  "prettier.singleQuote": true,
+  "prettier.jsxSingleQuote": true
 }
