+++ src/components/AppWrapper/index.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 import DeckGlMap from '../map';
 import Sidebar from '../Sidebar';
 import Nav from '../Nav';
-import Legend from '../Legend/Legend';
+import MapLayerLegend from '../Legend/MapLayersLegend';
 import Cookie from '../Cookie';
 import Loading from '../Loading';
 import Overlay from '../Overlay';
@@ -55,7 +55,7 @@ const Map: FC<{
 }> = ({ showOverlay, isNavOpened }) => {
   const data = useStoreState('data');
   const rainGeojson = useStoreState('rainGeojson');
-  const visibleLayer = useStoreState('visibleLayer');
+  const visibleMapLayer = useStoreState('visibleMapLayer');
   const pumps = useStoreState('pumps');
   const ageRange = useStoreState('ageRange');
   const dataView = useStoreState('dataView');
@@ -75,7 +75,7 @@ const Map: FC<{
       }}
       data={data || null}
       rainGeojson={rainGeojson || null}
-      visibleLayer={visibleLayer}
+      visibleMapLayer={visibleMapLayer}
       isTreeDataLoading={!!isTreeDataLoading}
       isNavOpen={!!isNavOpened}
       showControls={showOverlay}
@@ -121,7 +121,7 @@ const AppWrapper: FC = () => {
       <CookieContainer>
         <Cookie />
       </CookieContainer>
-      {showMapUI && <Legend />}
+      {showMapUI && <MapLayerLegend />}
       <ImprintAndPrivacyContainer></ImprintAndPrivacyContainer>
     </AppWrapperDiv>
   );
