--- src/components/Legend/FlexRow.tsx
@@ -1,13 +1,0 @@
-import styled from 'styled-components';
-
-export const FlexRow = styled.div`
-  display: flex;
-  flex-direction: colmun;
-  height: 16px;
-  line-height: 16px;
-  padding: 4px 9px;
-  width: 100%;
-  align-items: center;
-
-  margin-right: 10px;
-`;
