--- src/components/Legend/ItemLabel.tsx
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-export const ItemLabel = styled.label`
-  font-size: ${p => p.theme.fontSizeL};
-  opacity: 0.66;
-  padding-top: 5px;
-  width: 100%;
-`;
