--- src/components/Legend/PumpLabel.tsx
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-export const PumpLabel = styled.label`
-  font-size: ${p => p.theme.fontSizeL};
-  opacity: 0.66;
-  width: 100%;
-`;
