+++ src/components/Legend/RainColorLegend.tsx
@@ -1,18 +1,17 @@
 import React from 'react';
 import { interpolateColor } from '../../utils/colorUtil';
-import { legendLabels } from './legendLabels';
-import { ItemContainer } from './ItemContainer';
-import { LegendRect } from './LegendRect';
-import { ItemLabel } from './ItemLabel';
+import { ItemLabel, legendLabels } from './LegendLabels';
+import { LegendRect } from './LegendRectsDots';
+import { FlexColumnCentered } from './LegendFlexBoxes';
 
 export const RainColorLegend: React.FC = () => (
   <>
     {legendLabels.map(item => (
       <React.Fragment key={item.label}>
-        <ItemContainer>
+        <FlexColumnCentered>
           <LegendRect gradient={interpolateColor(item.value)} />
           <ItemLabel>{item.label}</ItemLabel>
-        </ItemContainer>
+        </FlexColumnCentered>
       </React.Fragment>
     ))}
   </>
