+++ src/components/Legend/TreesColorLegend.tsx
@@ -1,19 +1,18 @@
 import React from 'react';
 import { interpolateColor } from '../../utils/colorUtil';
-import { legendLabels } from './legendLabels';
-import { LegendDot } from './Legend';
-import { ItemContainer } from './ItemContainer';
-import { ItemLabel } from './ItemLabel';
+import { legendLabels, ItemLabel } from './LegendLabels';
+import { FlexColumnCentered } from './LegendFlexBoxes';
+import { LegendDot } from './LegendRectsDots';
 
 export const TreesColorLegend: React.FC = () => (
   <>
     {legendLabels.map(item => {
       return (
         <React.Fragment key={item.label}>
-          <ItemContainer>
+          <FlexColumnCentered>
             <LegendDot color={interpolateColor(item.value)} />
             <ItemLabel>{item.label}</ItemLabel>
-          </ItemContainer>
+          </FlexColumnCentered>
         </React.Fragment>
       );
     })}
