--- src/components/Legend/UnstyledFlex.tsx
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-import { Flex } from './Flex';
-
-export const UnstyledFlex = styled(Flex)`
-  border-bottom: none;
-  justify-content: flex-start;
-  margin-bottom: 0;
-`;
